requirements.txt
pip>=25.2
pyyaml>=6.0.2
#ruamel.yaml>=0.18.14
uptime_kuma_api